+++ zcash_client_sqlite/src/wallet/init/migrations/add_transaction_views.rs
@@ -25,7 +25,7 @@ const DEPENDENCIES: &[Uuid] = &[
 
 pub(crate) struct Migration;
 
-impl schemerz::Migration for Migration {
+impl schemerz::Migration<Uuid> for Migration {
     fn id(&self) -> Uuid {
         MIGRATION_ID
     }