+++ zcash_client_sqlite/src/wallet/init/migrations/fix_broken_commitment_trees.rs
@@ -20,7 +20,7 @@ pub(super) struct Migration<P> {
     pub(super) params: P,
 }
 
-impl<P> schemerz::Migration for Migration<P> {
+impl<P> schemerz::Migration<Uuid> for Migration<P> {
     fn id(&self) -> Uuid {
         MIGRATION_ID
     }