+++ zcash_client_sqlite/src/wallet/init/migrations/initial_setup.rs
@@ -11,7 +11,7 @@ pub(super) const MIGRATION_ID: Uuid = Uuid::from_u128(0xbc4f5e57_d600_4b6c_990f_
 
 pub(super) struct Migration;
 
-impl schemerz::Migration for Migration {
+impl schemerz::Migration<Uuid> for Migration {
     fn id(&self) -> Uuid {
         MIGRATION_ID
     }