+++ zcash_client_sqlite/src/wallet/init/migrations/sent_notes_to_internal.rs
@@ -15,7 +15,7 @@ const DEPENDENCIES: &[Uuid] = &[ufvk_support::MIGRATION_ID];
 
 pub(super) struct Migration;
 
-impl schemerz::Migration for Migration {
+impl schemerz::Migration<Uuid> for Migration {
     fn id(&self) -> Uuid {
         MIGRATION_ID
     }