+++ zcash_client_sqlite/src/wallet/init/migrations/v_transactions_shielding_balance.rs
@@ -16,7 +16,7 @@ const DEPENDENCIES: &[Uuid] = &[v_tx_outputs_use_legacy_false::MIGRATION_ID];
 
 pub(super) struct Migration;
 
-impl schemerz::Migration for Migration {
+impl schemerz::Migration<Uuid> for Migration {
     fn id(&self) -> Uuid {
         MIGRATION_ID
     }