+++ zcash_client_sqlite/src/wallet/init/migrations/v_transactions_transparent_history.rs
@@ -16,7 +16,7 @@ const DEPENDENCIES: &[Uuid] = &[sapling_memo_consistency::MIGRATION_ID];
 
 pub(super) struct Migration;
 
-impl schemerz::Migration for Migration {
+impl schemerz::Migration<Uuid> for Migration {
     fn id(&self) -> Uuid {
         MIGRATION_ID
     }